Module rfq::accounts

An Anchor generated module, providing a set of structs mirroring the structs deriving Accounts, where each field is a Pubkey. This is useful for specifying accounts for a client.

Structs

Confirm

Generated client accounts for Confirm.

Initialize

Generated client accounts for Initialize.

LastLook

Generated client accounts for LastLook.

Request

Generated client accounts for Request.

Respond

Generated client accounts for Respond.

ReturnCollateral

Generated client accounts for ReturnCollateral.

Settle

Generated client accounts for Settle.